+++ packaging/homebrew/shopify-cli.base.rb
@@ -98,6 +98,7 @@ class ShopifyCli < Formula
     (bin + file.basename).open("w") do |f|
       f << <<~RUBY
         #!#{ruby_bin}/ruby --disable-gems
+        ENV['ORIGINAL_ENV']=ENV.to_h.to_json
         ENV['GEM_HOME']="#{prefix}"
         ENV['GEM_PATH']="#{prefix}"
         ENV['RUBY_BINDIR']="#{ruby_bin}/"